refactor(pages): unify spawn into platform/, expose spawn_task from prelude
Merge spawn.rs into platform/{wasm,native}.rs and reduce platform.rs to
cfg dispatch. Expose spawn_task/defer_yield from reinhardt_pages::prelude
as cross-target API (no-op on native) and deprecate the direct
wasm_bindgen_futures::spawn_local re-export. Update internal call sites
and form! codegen to use crate::platform::spawn_task.

Refs #4362
Fixes #4365
