+++ crates/reinhardt-pages/src/lib.rs
@@ -180,9 +180,8 @@ pub mod callback;
 pub mod dom;
 pub mod logging;
 pub mod reactive;
-pub mod spawn;
 
-// Platform abstraction (unified types for WASM and native)
+// Platform abstraction (unified types and task spawning for WASM and native)
 pub mod platform;
 
 // Unified prelude for simplified imports
